+++ lib/components/features/explore/TopPerformers.tsx
@@ -6,7 +6,7 @@ import { TopTrades } from './TopTrades';
 
 export function TopPerformers() {
   return (
-    <div className="flex flex-col gap-4 relative">
+    <div className="flex flex-col gap-5 relative">
       <p className="text-xl font-semibold">Top Performers</p>
       <div className="flex flex-col md:flex-row gap-4">
         <TopVolume />